+++ apps/laboratory/src/utils/ConstantsUtil.ts
@@ -12,8 +12,9 @@ import {
   baseSepolia,
   gnosis,
   unichainSepolia,
-  type AppKitNetwork
+  hedera
 } from '@reown/appkit/networks'
+import type { AppKitNetwork } from '@reown/appkit/networks'
 import { getLocalStorageItem } from './LocalStorage'
 import type { CaipNetwork } from '@reown/appkit'
 
@@ -22,7 +23,7 @@ if (!projectId) {
   throw new Error('NEXT_PUBLIC_PROJECT_ID is not set')
 }
 export const WALLET_URL = process.env['WALLET_URL'] || 'https://react-wallet.walletconnect.com/'
-export const USEROP_BUILDER_SERVICE_BASE_URL = 'https://rpc.walletconnect.com/v1/wallet'
+export const USEROP_BUILDER_SERVICE_BASE_URL = 'https://rpc.walletconnect.org/v1/wallet'
 
 export const GALLERY_URL = 'https://appkit-gallery.reown.com/'
 export const DOCS_URL = 'https://docs.reown.com/appkit/overview'
@@ -62,7 +63,8 @@ const EvmNetworks = [
   baseSepolia,
   unichainSepolia,
   sepolia,
-  gnosis
+  gnosis,
+  hedera
 ] as [AppKitNetwork, ...AppKitNetwork[]]
 
 export const solanaNotExist = {