+++ packages/core/src/controllers/OptionsController.ts
@@ -10,6 +10,7 @@ import type {
 } from '../utils/TypeUtil.js'
 import { ConstantsUtil } from '../utils/ConstantsUtil.js'
 import { filterSocialsByPlatform } from '../utils/OptionsUtil.js'
+import type { SIWXConfig } from '../utils/SIWXUtil.js'
 
 // -- Types --------------------------------------------- //
 export interface OptionsControllerStatePublic {
@@ -111,6 +112,12 @@ export interface OptionsControllerStatePublic {
    * @see https://docs.reown.com/appkit/react/core/options#features
    */
   features?: Features
+  /**
+   * @experimental - This feature is not production ready.
+   * Enable Sign In With X (SIWX) feature in your AppKit.
+   * @default undefined
+   */
+  siwx?: SIWXConfig
 }
 
 export interface OptionsControllerStateInternal {
@@ -153,10 +160,12 @@ export const OptionsController = {
       if (!state.features) {
         state.features = ConstantsUtil.DEFAULT_FEATURES
       }
+
       let filteredValue = value
       if (key === 'socials') {
         filteredValue = filterSocialsByPlatform(state.features.socials)
       }
+
       if (key in state.features) {
         ;(state.features as Record<keyof Features, unknown>)[key as keyof Features] = filteredValue
       }
@@ -237,5 +246,9 @@ export const OptionsController = {
 
   setHasMultipleAddresses(hasMultipleAddresses: OptionsControllerState['hasMultipleAddresses']) {
     state.hasMultipleAddresses = hasMultipleAddresses
+  },
+
+  setSIWX(siwx: OptionsControllerState['siwx']) {
+    state.siwx = siwx
   }
 }