+++ packages/experimental/src/smart-session/ui/permissions/wui-permission-contract-call/index.ts
@@ -19,13 +19,13 @@ export class WuiPermissionContractCall extends LitElement {
     }
 
     return html`
-      <wui-flex flexDirection="column" padding=${['l', '0', '0', '0']} alignItems="center">
+      <wui-flex flexDirection="column" alignItems="center">
         <wui-details-group>
           <wui-details-group-item name="Type">
-            <wui-text variant="paragraph-400" color="fg-100"> Contract Call </wui-text>
+            <wui-text variant="small-400" color="fg-100"> Contract Call </wui-text>
           </wui-details-group-item>
           <wui-details-group-item name="Contract">
-            <wui-text variant="paragraph-400" color="fg-100">
+            <wui-text variant="small-400" color="fg-100">
               ${UiHelperUtil.getTruncateString({
                 string: this.contractAddress,
                 truncate: 'middle',
@@ -35,14 +35,14 @@ export class WuiPermissionContractCall extends LitElement {
             </wui-text>
           </wui-details-group-item>
           <wui-details-group-item name="Functions">
-            <wui-text variant="paragraph-400" color="fg-100">
+            <wui-text variant="small-400" color="fg-100">
               ${this.functions?.map(f => f.functionName).join(', ')}
             </wui-text>
           </wui-details-group-item>
           <wui-flex justifyContent="space-between">
             <wui-text color="fg-200">Duration</wui-text>
             <wui-flex flexDirection="column" alignItems="flex-end" gap="s">
-              <wui-text variant="paragraph-400" color="fg-100">
+              <wui-text variant="small-400" color="fg-100">
                 ~ ${Math.round((1000 * this.expiry - Date.now()) / 1000 / 3600)} hours
               </wui-text>
               <wui-text variant="tiny-600" color="fg-300">